checkpoint on refactoring
