+++ lib/coverband.rb
@@ -1,5 +1,6 @@
 require 'redis'
 require 'logger'
+# todo this shouldn't be required, but allowed if they use S3 output
 require 'aws-sdk'
 
 require 'coverband/version'
@@ -8,7 +9,9 @@ require 'coverband/adapters/redis_store'
 require 'coverband/adapters/memory_cache_store'
 require 'coverband/base'
 require 'coverband/baseline'
+require 'coverband/reporters/base'
 require 'coverband/reporters/simple_cov_report'
+require 'coverband/reporters/console_report'
 require 'coverband/middleware'
 require 'coverband/s3_report_writer'
 
@@ -20,12 +23,9 @@ module Coverband
     attr_accessor :configuration_data
   end
 
+  # this method is left for backwards compatibility with existing configs
   def self.parse_baseline(baseline_file = './tmp/coverband_baseline.json')
-    baseline = if File.exist?(baseline_file)
-      JSON.parse(File.read(baseline_file))
-    else
-      {}
-    end
+    Coverband::Baseline.parse_baseline(baseline_file)
   end
 
   def self.configure(file = nil)
