+++ test/unit/adapters_memory_cache_store_test.rb
