+++ test/unit/adapters_redis_store_test.rb
