+++ test/unit/reports_simple_cov_test.rb
@@ -1,0 +1,35 @@
+require File.expand_path('../test_helper', File.dirname(__FILE__))
+
+class ReportsSimpleCovTest < Test::Unit::TestCase
+
+  def setup
+    @fake_redis = fake_redis
+    @store = Coverband::Adapters::RedisStore.new(@fake_redis, array: true)
+  end
+
+  test "report data with scov" do
+    Coverband.configure do |config|
+      config.redis             = @fake_redis
+      config.reporter          = 'scov'
+      config.s3_bucket         = nil
+    end
+
+    Coverband::Reporters::SimpleCovReport.expects(:current_root).at_least_once.returns('/root_dir')
+    fake_redis.expects(:smembers).with('coverband').returns(fake_coverband_members)
+
+    SimpleCov.expects(:track_files)
+    SimpleCov.expects(:add_not_loaded_files).returns({})
+    SimpleCov::Result.any_instance.expects(:format!)
+
+    fake_coverband_members.each do |key|
+      File.expects(:exists?).with(key).returns(true)
+      File.expects(:foreach).with(key).returns(['a','b','c'])
+      fake_redis.expects(:smembers).with("coverband.#{key}").returns(["54", "55"])
+    end
+    
+    Coverband.configuration.logger.stubs('info')
+
+    Coverband::Reporters::SimpleCovReport.report(@store, open_report: false)
+  end
+
+end
